Exception handling

+++ packages/api/resolvers/mutations/removeProductReview.js
@@ -1,6 +1,6 @@
 import { log } from 'meteor/unchained:core-logger';
 import { ProductReviews } from 'meteor/unchained:core-products';
-import { ProductReviewNotFoundError } from '../../errors';
+import { ProductReviewNotFoundError, InvalidIdError } from '../../errors';
 
 export default function removeProductReview(
   root,
@@ -8,6 +8,7 @@ export default function removeProductReview(
   { userId },
 ) {
   log('mutation removeProductReview', { userId, productReviewId });
+  if (!productReviewId) throw new InvalidIdError({ productReviewId });
   if (ProductReviews.find({ _id: productReviewId }).count() === 0)
     throw new ProductReviewNotFoundError({ productReviewId });
   return ProductReviews.deleteReview({
